Update named instances branch with latest exploratory code.
I haven't updated the branch in several days despite a lot of work being done. Unfortunately, the work done has been mostly poking and proding the existing code to understand how the instance id mappings work.

Currently the reserve_named_instances function calls a new generate_unique_id_from_name function that is simply returning the length of the name. This is for testing purposes.
The set_named_inst_value function also treats the name->id conversion as simply the length. Again, this will be updated to be significantly more robust.

The test being ran right now is not a full fledged test, but rather simply a mechanism for me to see what is taking place in the system.